Working intensely on I/O
Previously implemented a rudimentary storage peripheral for the emulator, but now that work has begun towards building a whole computer, serious consideration needs to be given to the peripherals, I/O, and interrupt system.

Started experimenting with X11 and curses, and for now, will pursue further work with curses for emulated terminal. Will begin with research and rework of memory-mapped I/O with consideration given for the primitive interrupt system on the 8080.