Implemented I/O State Machines
In real life, I/O devices would be controlled by state machines in the form of digital logic circuits. So in order to solve the problem of implementing I/O devices in the emulator, it made sense to turn to state machines.

Both the storage and keyboard are now implemented using rudimentary state machines. They have a clear set of possible states that can include
	-READY
	-READING
	-WRITING
	-INTERRUPT
	-OP_COMPLETE
and a clear set of inputs that can include
	-INTERRUPT_ENABLE
	-RDY
	-READ_REQUEST
	-WRITE_REQUEST
	-DONE