Modified Assembler to exit ONLY through main
Previously, the assembler was written so that if mallocs in side of procedure calls failed, the assembler would exit right then and there. This is a problem because it doesn't follow good coding style, which requires that all dynamically-allocated memory be cleared prior to exit.

In the assembler, there are MANY dynamically-allocated structures, so in order to properly free all that memory prior to exit, all malloc-ing procedure calls were modified to return a success or failure value. Then in main, this value can be processed to continue with the program or free the allocated memory and exit.