Starting address at beginning of every object code
The object code that is produced by the assembler is based on the output of other assemblers found online, and is structured as follows:

Number of bytes (n) from current starting address (1 Byte)
Current starting address (2 Bytes)
machine code (n bytes)

This structure repeats every time an ORG directive shows up in the assembly code.
The problem is that when a programmer writes code, they usually assume the very first starting address to be 0x0000, and omit an ORIG for that address. This would cause an error when the object code is interpretted by a loader in the computer, which always expects the number of bytes, and the current starting address, prior to the actual code.

To deal with this problem, the buffer that takes in the assembly code from the assembly code file is instantiated with "ORG 0000\n" so that a default starting location is generated in the object code in order to prevent errors during program loading.