Fixed Processing of "EQU" Pseudo-instruction
The assembler was having problems with generating correct object code for programs that contained a lot of labels produced by "EQU". A closer investigation of the problem found two problems.

The first was the instantiation of the b1 buffer was creating labels that contained a " " at the very beginning. This is a problem because when the label is then compared against an operand input, any labels in the operand won't match with label containing a " ".

The second problem was the unchecked case where there was an identified label was--for example--"abed", and the intend label was "evil_abed".

The first problem was solved by simply using the "ShiftBufferContentsLeft" function once.
The second problem was solved by check the character--in the operand--that came before identified label. If it was a " " or a tab, it indicated that this in fact was the correct label, and if it was not, it indicated that that the assembler should continue looking.