Implemented the Define-Byte (DB) and Define-Word (DW) Pseudo-Instructions
Most programs contain not only instructions but also data. But up to this point, the assembler did not provide the programmer the capability to store data.

In order to fix this, the well-known Define-Byte and Define-Word pseudo-instructions were implemented. These instructions allow data to be stored in the code, and allow for the data to be easily accessed using labels.