Fixed byte_counter
Previously byte_counter failed to generate and store the correct number of bytes--for a given starting address--to the object code. This was because the program was simply adding up the number of instructions between the current ORG and the next ORG or end-of-code.

In order to fix this problem, the simple incrementing was replaced by a switch statement that takes into account the type of instruction in order to add the proper number of bytes to the byte_counter.

In addition to this change, there was further investigation into the function of an existing 8080 assembler, and it was discovered that the byte count in those assemblers only goes up to 0x10 before generating another byte-count and address.

For now, this assembler will keep the existing code, and expand the number of bytes given to the byte-count from 1 to 2.

In the future, the byte-counting scheme may change.