Modified GetProgram Function to Correctly Process Object Code Generated by the Assembler
GetProgram was originally written to take in only bytes of data that would comprise a program.
This would not allow for the byte-count and starting address* that is also included in object code generated by the assembler.

In order to fix this, GetProgram was modified to take in a byte-count and starting address whenever an local variable byte_count was 0.

*- The starting address is the address from which the following series of n bytes (in the object code) is placed, where n = byte-count.